% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_grid.R
\name{paged_grid}
\alias{paged_grid}
\title{Function for grid template}
\usage{
paged_grid(logo = "0", other_css = NULL, ...)
}
\arguments{
\item{logo}{Logo}

\item{other_css}{Add an other CSS}

\item{...}{Arguments passed to pagedown::html_paged}
}
\value{
A pagedown template
}
\description{
Function for grid template
}
