% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_hosting.R
\name{paged_hosting}
\alias{paged_hosting}
\title{Function for hosting template}
\usage{
paged_hosting(
  logo = "0",
  front_img = "0",
  back_img = "0",
  img_to_dark = TRUE,
  logo_to_white = FALSE,
  other_css = NULL,
  toc = TRUE,
  toc_depth = 2,
  ...
)
}
\arguments{
\item{logo}{Logo}

\item{front_img}{Front cover image}

\item{back_img}{Back cover image}

\item{img_to_dark}{Darken image}

\item{logo_to_white}{Tranform logo color to white - logo must be SVG}

\item{other_css}{Add an other CSS}

\item{toc}{Table of contents}

\item{toc_depth}{TOC depth}

\item{...}{Arguments passed to pagedown::html_paged}
}
\value{
A pagedown template
}
\description{
Function for hosting template
}
