% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_mine.R
\name{paged_mine}
\alias{paged_mine}
\title{Function for my template}
\usage{
paged_mine(
  logo = "0",
  front_img = "0",
  img_to_dark = TRUE,
  logo_to_white = FALSE,
  other_css = NULL,
  ...
)
}
\arguments{
\item{logo}{Logo}

\item{front_img}{Front image}

\item{img_to_dark}{Darken image}

\item{logo_to_white}{Tranform logo color to white - logo must be SVG}

\item{other_css}{Add an other CSS}

\item{...}{Arguments passed to pagedown::html_paged}
}
\value{
A pagedown template
}
\description{
Function for my template
}
