% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_morant.R
\name{paged_morant}
\alias{paged_morant}
\title{Function for Morant template}
\usage{
paged_morant(logo = "0", other_css = NULL, ...)
}
\arguments{
\item{logo}{Logo}

\item{other_css}{Add an other CSS}

\item{...}{Arguments passed to pagedown::html_paged}
}
\value{
A pagedown template
}
\description{
Function for Morant template
}
