% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_reportcard.R
\name{paged_reportcard}
\alias{paged_reportcard}
\title{Function for take3 template}
\usage{
paged_reportcard(
  logo = "Kids",
  front_img = "Kids",
  img_to_dark = TRUE,
  logo_to_white = FALSE,
  other_css = NULL,
  ...
)
}
\arguments{
\item{logo}{Logo}

\item{front_img}{Front cover image}

\item{img_to_dark}{Darken image}

\item{logo_to_white}{Transform logo color to white - logo must be SVG}

\item{other_css}{Add an other CSS}

\item{...}{Arguments passed to pagedown::html_paged}
}
\value{
A pagedown template
}
\description{
Function for take3 template
}
