% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_windmill.R
\name{paged_windmill}
\alias{paged_windmill}
\title{Function for windmill template}
\usage{
paged_windmill(logo = "0", img = "0")
}
\arguments{
\item{logo}{Logo}

\item{img}{Image}
}
\value{
A pagedown template
}
\description{
Function for windmill template
}
