% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_grid.R
\name{paged_grid}
\alias{paged_grid}
\title{Function for grid template}
\usage{
paged_grid(logo = "0")
}
\arguments{
\item{logo}{Logo}
}
\value{
A pagedown template
}
\description{
Function for grid template
}
