% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged_windmill.R
\name{paged_windmill}
\alias{paged_windmill}
\title{Function for windmill template}
\usage{
paged_windmill(
  logo = "0",
  img = "0",
  img_to_dark = TRUE,
  logo_to_white = FALSE
)
}
\arguments{
\item{logo}{Logo}

\item{img}{Image}

\item{img_to_dark}{Darken image}

\item{logo_to_white}{Tranform logo color to white}
}
\value{
A pagedown template
}
\description{
Function for windmill template
}
